% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_method.R
\name{append_method}
\alias{append_method}
\title{To attach and compare new method to existing comparison study done with function 'impute_errors()'}
\usage{
append_method(existing_method, dataIn, missPercentFrom, missPercentTo, interval,
  repetition, errorParameter, MethodPath, MethodName)
}
\arguments{
\item{existing_method}{as Error observations for different methods}

\item{dataIn}{as imput time series for testing}

\item{missPercentFrom}{as variable from which percent of missing values to be considered}

\item{missPercentTo}{as variable to state upto what percent missing values are to be considered}

\item{interval}{as interval between consecutive missPercent values}

\item{repetition}{as an integer to state number of repetition to be done for each missPercent value}

\item{errorParameter}{as type of error calculation (RMSE, MAE or MAPE)}

\item{MethodPath}{as location of function for proposed imputation method}

\item{MethodName}{as name for function for proposed imputation method}
}
\value{
Returns error comparosin for imputation methods
}
\description{
To attach and compare new method to existing comparison study done with function 'impute_errors()'
}
\examples{
#Kindly, refer "Vignette" document
}

