% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_functions.R
\name{mae}
\alias{mae}
\title{Mean Absolute Error Calculation}
\usage{
mae(diff)
}
\arguments{
\item{diff}{is the difference between Original data and Predicted data}
}
\value{
maeVal as Mean Absolute Error
}
\description{
takes difference between Original data and Predicted data as input
}
\examples{
## Generate 100 random numbers within some limits
x <- sample(1:7, 100, replace = TRUE)
y <- sample(1:4, 100, replace = TRUE)
z <- mae(x - y)
z
}

