% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_errors.R
\name{plot_errors}
\alias{plot_errors}
\alias{plot_errors.errprof}
\title{Function to plot the Error Comparison}
\usage{
plot_errors(dataIn, ...)

\method{plot_errors}{errprof}(dataIn, plotType = c("boxplot", "bar", "line"),
  ...)
}
\arguments{
\item{dataIn}{an errprof object returned from \code{\link{inpute_errors}}}

\item{plotType}{chr string indicating plot type}

\item{\dots}{arguments passed to or from other methods}
}
\value{
a ggplot object that can be further modified
}
\description{
Function to plot the Error Comparison
}
\examples{
aa <- impute_errors()
plot_errors(aa)

cc <- impute_errors()
plot_errors(dataIn = cc, plotType = 1, main = "Bar plot",
         args.legend = list(x="topleft", bg = "NA"))
plot_errors(dataIn = cc, plotType = 2)
}

