% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/print.errprof.R
\name{print.errprof}
\alias{print.errprof}
\title{Print method for errprof}
\usage{
\method{print}{errprof}(x, ...)
}
\arguments{
\item{x}{input errprof object}

\item{...}{arguments passed to or from other methods}
}
\value{
list output for the errprof object
}
\description{
Print method for errprof class
}

