% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_method.R
\name{remove_method}
\alias{remove_method}
\title{Removes unwanted method from already existing methods in Test bench}
\usage{
remove_method(existing_method, index_number)
}
\arguments{
\item{existing_method}{as Error observations for different methods}

\item{index_number}{as index number of unwanted method in study}
}
\value{
it removes unwanted method from test bench and returns other method errors
}
\description{
Removes unwanted method from already existing methods in Test bench
}

