% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_errors.R
\name{impute_errors}
\alias{impute_errors}
\title{Function working as testbench for comparison of imputing models}
\usage{
impute_errors(dataIn = NULL, smps = "mcar", blck = 0.5, blckper = TRUE,
  missPercentFrom = 10, missPercentTo = 90, interval = 10,
  repetition = 10, errorParameter = "rmse", MethodPath = NULL,
  MethodName = "Proposed Method")
}
\arguments{
\item{dataIn}{input \code{\link[stats]{ts}} for testing, defaults to \code{\link[datasets]{nottem}}}

\item{smps}{chr string indicating sampling type for generating missing data, see details}

\item{blck}{numeric indicating block sizes as a proportion of the sample size for the missing data, applies only if \code{smps = 'mcar'}}

\item{blckper}{logical indicating if the value passed to \code{blck} is a proportion of the sample size for missing data, otherwise \code{blck} indicates number of observations}

\item{missPercentFrom}{numeric from which percent of missing values to be considered}

\item{missPercentTo}{numeric for up to what percent missing values are to be considered}

\item{interval}{numeric for interval between consecutive missPercent values}

\item{repetition}{numeric for repetitions to be done for each missPercent value}

\item{errorParameter}{chr string indicating which error type to use, acceptable values are \code{"rmse"} (default), \code{"mae"}, \code{"mape"}, or \code{"other"}}

\item{MethodPath}{chr string of location of function for the proposed imputation method}

\item{MethodName}{chr string for name for function for the proposed imputation method}
}
\value{
Returns error comparison for imputation methods
}
\description{
Function working as testbench for comparison of imputing models
}
\details{
The \code{smps} argument indicates the type of sampling for generating missing data.  Options are \code{smps = 'mcar'} for missing completely at random and \code{smps = 'mar'} for missing at random.  Additional information about the sampling method is described in \code{\link{sample_dat}}.
}
\examples{
aa <- impute_errors()
aa
}
\seealso{
\code{\link{sample_dat}}
}

