% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_errors.R
\name{impute_errors}
\alias{impute_errors}
\title{Function working as testbench for comparison of imputing models}
\usage{
impute_errors(dataIn, missPercentFrom, missPercentTo, interval, repetition,
  errorParameter, MethodPath, MethodName, random, startPoint, patchLength)
}
\arguments{
\item{dataIn}{as input data for testing}

\item{missPercentFrom}{as variable from which percent of missing values to be considered}

\item{missPercentTo}{as variable to state upto what percent missing values are to be considered}

\item{interval}{as interval between consecutive missPercent values}

\item{repetition}{as an integer to decide the numbers of repetition to be done for each missPercent value}

\item{errorParameter}{as type of error calculation (RMSE, MAE or MAPE)}

\item{MethodPath}{as location of function for the proposed imputation method}

\item{MethodName}{as name for function for the proposed imputation method}

\item{random}{as parameter to decide missing value scheme}

\item{startPoint}{as index of missing patches in dataset when 'random = 0'}

\item{patchLength}{as length of missing patches in dataset when 'random = 0'}
}
\value{
Returns error comparison for imputation methods
}
\description{
Function working as testbench for comparison of imputing models
}
\examples{
aa <- impute_errors()
aa

bb <- impute_errors(random = 0, startPoint = c(10, 20, 40), patchLength = c(3, 4, 5))
bb
\dontrun{
cc <- impute_errors()
cc

dd <- impute_errors(random = 0, startPoint = c(1,10,30,0,3,50,70,90), patchLength = c(1,5,5,0,3,10,10,10))
dd
}
}

