% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_errors.R
\name{plot_errors}
\alias{plot_errors}
\title{Function to plot the Error Comparison}
\usage{
plot_errors(dataIn, plotType, ...)
}
\arguments{
\item{dataIn}{as input data in list format (returned by function inpute_errors())}

\item{plotType}{as parameter to decide Plot Type to achieve: 1: Barplot, 2: Line plot}

\item{\dots}{as additive functions for plot}
}
\value{
It returns the Error comparison for different methods
}
\description{
Function to plot the Error Comparison
}
\examples{
# aa <- impute_errors()
# bb <- plot_errors(aa)
# bb

# cc <- impute_errors()
# dd <- plot_errors(dataIn = cc, plotType = 1, main = "Bar plot",
#         args.legend = list(x="topleft", bg = "NA"))

# cc <- impute_errors()
# dd <- plot_errors(dataIn = cc, plotType = 2)
}

