% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_errors.R
\name{plot_errors}
\alias{plot_errors}
\title{Function to plot the Error Comparison}
\usage{
plot_errors(dataIn)
}
\arguments{
\item{dataIn}{as input data in list format (returned by function inpute_errors())}
}
\value{
It returns the Error comparison for different methods
}
\description{
Function to plot the Error Comparison
}
\examples{
# aa <- impute_errors()
# bb <- plot_errors(aa)
# bb
}

