% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_functions.R
\name{rmse}
\alias{rmse}
\title{Root Mean Square Error Calculation}
\usage{
rmse(obs, pred)
}
\arguments{
\item{obs}{numeric vector of original data}

\item{pred}{numeric vector of predicted data}
}
\value{
rmseVal as Root Mean Square Error
}
\description{
takes difference between Original data and Predicted data as input
}
\examples{
## Generate 100 random numbers within some limits
x <- sample(1:7, 100, replace = TRUE)
y <- sample(1:4, 100, replace = TRUE)
z <- rmse(x, y)
z
}

